requirements.txt
requests
pandas
python-dateutil
psycopg2-binary
python-dotenv